CELERY WORKER BEAT

+++ accounts/tasks.py
@@ -1,6 +1,9 @@
 from celery import shared_task
 from django.utils import timezone
 from .models import OTPVerification
+import logging
+
+logger = logging.getLogger(__name__)
 
 @shared_task
 def delete_expired_otps():
@@ -9,5 +12,7 @@ def delete_expired_otps():
         created_at__lt=timezone.now() - timezone.timedelta(minutes=5)
     )
     count = expired.count()
-    expired.delete()
+    if count > 0:
+        expired.delete()
+        logger.info(f"{count} OTP удалено")
     return f"{count} OTP удалено"
